CFIN-292 refactor page head to separate component
